ndk: Add missing API levels, use intra-links in AAudio docs (#108)
* ndk: Add missing API levels

A bunch of AAudio functions use api-level-28 but this feature did not
exist.

For convenience add the missing 27 and 25 as well, in case new wrappers
are introduced that may also carry a bunch of functions introduced on
these levels. Now the supported list is linear.

* aaudio: Add back data_callback docs and convert `` to intra-links

* CI: Check all features for the NDK

* CI: Build NDK docs and disable warnings

This prevents ie. broken_intra_doc_links from ending up in the main
branch.

* ndk: Convert links to autolinks

Nightly rustdoc warns/errors that these links are not converted to
hyperlinks (and thus not clickable in the web pages).

Note that the link to AudioTrack#write is broken; rustdoc rightfully
converts `[]` to `%5B%5D`, but android.com does not resolve that to the
right anchor.